yolo_obb_editor_web/requirements.txt
streamlit>=1.30.0
streamlit-drawable-canvas-fix>=0.9.8
Pillow>=10.0.0
